put shared code in MouseableGLWindow.java
